Step8/requirements.txt
pytest==6.2.5
beautifulsoup4==4.10.0
requests==2.26.0
azure-storage-blob==12.12.0
mysql-connector-python==8.0.29
